Moved router to app.js and edit achievements
