+++ assets/data/achievements.js
@@ -2,46 +2,31 @@ export default [
     {
         id: '0',
         achievementName: 'Split a bill',
-        image: "https://cdn-icons.flaticon.com/png/512/4755/premium/4755473.png?token=exp=1649290873~hmac=cba56593c8a19c886a5c554df862e6d7"
+        image: require('../images/splitBill.png')
     },
     {
         id: '1',
         achievementName: 'Split 5 bills',
-        image: "https://cdn-icons.flaticon.com/png/512/4755/premium/4755504.png?token=exp=1649289974~hmac=6f92563e44c5718a4311eff278d16027"
+        image: require('../images/split5.png')
     },
     {
         id: '2',
         achievementName: 'On-time payment',
-        image: "https://cdn-icons.flaticon.com/png/512/4755/premium/4755524.png?token=exp=1649290967~hmac=2e9c2220beb77739c1bb716fba16f503"
+        image: require('../images/ontime.png')
     },
     {
         id: '3',
         achievementName: '5 consecutive on-time payments',
-        image: "https://cdn-icons.flaticon.com/png/512/4472/premium/4472415.png?token=exp=1649289401~hmac=73b6e053ee82b1dfd09febee43b0ed47"
+        image: require('../images/ontime5.png')
     },
     {
         id: '4',
         achievementName: 'First in a Leaderboard',
-        image: "https://cdn-icons.flaticon.com/png/512/6045/premium/6045678.png?token=exp=1649289802~hmac=63d23ecc1011d6066476a9dcd10d962b"
+        image: require('../images/first.png')
     },
     {
         id: '5',
-        achievementName: 'First in a Leaderboard for a week',
-        image: "https://cdn-icons.flaticon.com/png/512/6045/premium/6045974.png?token=exp=1649289802~hmac=4cba18b8a95c806122cd5798d104571e"
-    },
-    {
-        id: '6',
-        achievementName: 'First in a Leaderboard for a month',
-        image: "https://cdn-icons.flaticon.com/png/512/4755/premium/4755480.png?token=exp=1649290873~hmac=2fa7e9abdeedec2a0fe227425a039e83"
-    },
-    {
-        id: '7',
         achievementName: 'Invite a friend',
-        image: "https://cdn-icons.flaticon.com/png/512/4755/premium/4755486.png?token=exp=1649290024~hmac=0f2bf6a892f49b25cff07231de9b8fcb"
-    },
-    {
-        id: '8',
-        achievementName: 'Invite 5 friends',
-        image: "https://cdn-icons.flaticon.com/png/512/4472/premium/4472400.png?token=exp=1649289401~hmac=f24f0dddfb79408b1d5f74ad40a063ce"
+        image: require('../images/friends.png')
     },
 ]
