+++ src/components/Achievement/index.js
@@ -5,26 +5,33 @@ const AchievementsComponent = (props) => {
     const achievement = props.achievement;
     return (
         <View style={styles.container}>
+            <Image source={achievement.image} style={styles.image} />
             <Text style={styles.text}>{achievement.achievementName}</Text>
-            <Image source={{ uri: achievement.image }} style={styles.image} />
         </View>
     );
 };
 
 const styles = StyleSheet.create({
     container: {
-        height: 175,
-        width: 175,
+        height: 70,
+        width: '95%',
         backgroundColor: '#f6cfd2',
         borderRadius: 10,
         margin: 5,
         alignItems: 'center',
+        justifyContent: 'center',
+        flexDirection: 'row',
+        marginLeft: 10,
     },
     text: {
         color: 'grey',
         fontSize: 20,
         textAlign: 'center',
         padding: 10,
+        width: '70%'
+    },
+    imageContainer: {
+        //justifyContent: 'center'
     },
     image: {
         height: 50,
